refactor: updated directory structure to improve scalability
