+++ apps/HeadUnit/main.qml
@@ -1,14 +1,17 @@
 import QtQuick 2.15
-import QtQuick.Controls 2.15
 import QtQuick.Layouts
 
 
 Window {
     visible: true
-    visibility: "FullScreen"
+    // visibility: "FullScreen"
+    width: 800
+    height: 800
+
     title: qsTr("Head Unit")
     color: "gray"
     // todo background design
+    property Item previousItem: null
 
     ColumnLayout {
         anchors.fill: parent
@@ -156,7 +159,7 @@ Window {
                     color: "transparent"
 
                     Image {
-                        source: "qrc:/icons/menus"
+                        source: "qrc:/icons/menu"
                         width: 50
                         height: 50
                         anchors.centerIn: parent
@@ -166,7 +169,8 @@ Window {
                     MouseArea {
                         anchors.fill: parent
                         onClicked: {
-                            contentLoader.source = "Menu.qml"
+
+                            contentLoader.source = "modules/menu/menu.qml"
 
                             console.log("menu clicked")
                         }
@@ -179,14 +183,40 @@ Window {
                 Layout.fillWidth: true
                 Layout.fillHeight: true
                 Layout.rightMargin: 10
+                Text {
+                    id: name
+                    text: qsTr("text")
+                }
 
+
+                color: "transparent"
                 Loader {
                     id: contentLoader
                     anchors.fill: parent
-                    source: "Home.qml"
+                    source: "home.qml"
+                    // onItemChanged: {
+                    //     if (contentLoader.item) {
+
+                    //         console.log("새로운 컴포넌트 로드됨:", contentLoader.item);
+                    //         console.log("previousItem:", previousItem);
+                    //         console.log("contentLoader.item:", contentLoader.item);
+
+                    //         // 이전 컴포넌트가 존재하면 파괴 (혹은 메모리 정리 작업)
+                    //         if (contentLoader.item !== previousItem) {
+
+                    //             console.log("다르잔아");
+                    //             if (previousItem) {
+                    //                 previousItem.destroy();  // 이전 컴포넌트를 명시적으로 파괴
+                    //                 console.log("파괴ㅣㅣㅣ");
+                    //             }
+                    //               // 현재 아이템을 이전 아이템으로 설정
+                    //             previousItem = contentLoader.item;
+                    //         }
+                    //     }
+                    // }
                 }
             }
-
+        }
         }
     }
-}
+
