+++ apps/HeadUnit/shared/utils/envmanager.cpp
