+++ apps/HeadUnit/shared/utils/envmanager.h
