+++ apps/HeadUnit/shared/utils/utils.cpp
