+++ apps/HeadUnit/shared/utils/utils.h
